Made the grid change state when clicked
If a dead cell is clicked with the mouse, it becomes alive. Added two functions: One to check whether the area the mouse clicked is actually on the grid; another one to change the cell from dead to alive after that area has been clicked.
Added a randomizer as well that will pre fill the cells